refactor: 一旦 Home フォルダの解体
