+++ src/components/SearchCondition.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import { Tag, TagLabel, TagCloseButton, HStack } from "@chakra-ui/react";
 import { useHistory } from "react-router-dom";
-import { useAppDispatch } from "../../hooks/hooks";
+// XXX: 絶対パスで指定できていない
+import { useAppDispatch } from "../hooks/hooks";
 
 type Props = {
   searchCondition: string;
